+++ CHANGELOG.md
@@ -6,3 +6,4 @@
   - Setting `:default_exception_hander` Symbol to `I18n.exception_handler`.
   - `normalize_translation_keys` in favor of `normalize_keys`.
   - `:rescue_format` option on the exception handler.
+  - `enforce_available_locales` now defaults to true with no deprecation message.